Rename package to make it tech agnostic
