+++ movies_etl/config/config.yaml
@@ -4,8 +4,8 @@ dynaconf_merge:
 
 global:
   task_argument_class_mapping:
-    standardize: spark_movies_etl.tasks.StandardizeDataTask
-    curate: spark_movies_etl.tasks.CurateDataTask
+    standardize: movies_etl.tasks.StandardizeDataTask
+    curate: movies_etl.tasks.CurateDataTask
   movie_languages_filter: ["EN", "FR", "DE", "ES"]
 
 development:
