+++ movies_etl/tasks/__init__.py
@@ -1,0 +1,4 @@
+from movies_etl.tasks.curate_data.task import CurateDataTask
+from movies_etl.tasks.standardize_data.task import StandardizeDataTask
+
+__all__ = ["CurateDataTask", "StandardizeDataTask"]
