--- spark_movies_etl/tasks/__init__.py
@@ -1,4 +1,0 @@
-from spark_movies_etl.tasks.curate_data.task import CurateDataTask
-from spark_movies_etl.tasks.standardize_data.task import StandardizeDataTask
-
-__all__ = ["CurateDataTask", "StandardizeDataTask"]
