+++ tests/movies_etl/integration/fixtures/data-lake-raw-test/movie_ratings/2021/06/03/part-00000-e8b39350-08bf-41ff-9ad1-6b6c3b1922a0-c000.snappy.parquet
