+++ tests/movies_etl/unit/tasks/standardize_data/transformation_test.py
@@ -1,9 +1,13 @@
+from datetime import date
+
 from pyspark.sql import SparkSession
 
-from spark_movies_etl.schema import Schema
-from spark_movies_etl.tasks.curate_data.transformation import CurateDataTransformation
+from movies_etl.schema import Schema
+from movies_etl.tasks.standardize_data.transformation import (
+    StandardizeDataTransformation,
+)
 from tests.conftest import assert_data_frames_equal
-from tests.spark_movies_etl.unit.tasks.curate_data.fixtures.data import (
+from tests.movies_etl.unit.tasks.standardize_data.fixtures.data import (
     TEST_TRANSFORM_INPUT,
     TEST_TRANSFORM_OUTPUT_EXPECTED,
 )
@@ -11,14 +15,14 @@ from tests.spark_movies_etl.unit.tasks.curate_data.fixtures.data import (
 
 def test_transform(spark: SparkSession) -> None:
     # GIVEN
-    transformation = CurateDataTransformation(movie_languages=["EN", "ES", "DE", "FR"])
+    transformation = StandardizeDataTransformation(execution_date=date(2021, 1, 1))
     df_input = spark.createDataFrame(
         TEST_TRANSFORM_INPUT,  # type: ignore
-        schema=Schema.STANDARDIZED,
+        schema=Schema.RAW,
     )
     df_expected = spark.createDataFrame(
         TEST_TRANSFORM_OUTPUT_EXPECTED,  # type: ignore
-        schema=Schema.CURATED,
+        schema=Schema.STANDARDIZED,
     )
 
     # WHEN
